# Proof of concept: sjoin speedup with pygeos