I did a _very_ rough edit to `sjoin` to add `pygeos` support. I'm testing with the [ `pygeos` master branch](/pygeos/pygeos/tree/master) (not pypi package). I'm not using pygeos/pygeos#108, I expect the speedup to be much greater with that, which would be changing just 2 lines in `sjoin.py`.

Here are results (test script below):
| data_len | this branch + pygeos@master | this branch (no pygeos) | joris/pygeos + pygeos@master | joris/pygeos no pygeos | geopandas@master |
|----------|-----------------------------|-------------------------|------------------------------|------------------------|------------------|
| 1        | 0.200239                    | 0.206183                | 0.206509                     | 0.286895               | 0.269579         |
| 10       | 0.125859                    | 0.174937                | 0.286633                     | 0.204224               | 0.224107         |
| 100      | 0.152146                    | 0.33471                 | 0.73879                      | 0.518237               | 0.425057         |
| 1000     | 1.021618                    | 2.941023                | 34.455022                    | 2.938148               | 2.967173         |

In reality, I think a refactor of `sindex.py` and `sjoin.py` are needed to make this nice. At the very least, I would move the logic for picking a spatial indexing backend to `sindex.py` and I would split up `sjoin` similar to geopandas/geopandas#1271 so we can have separate functions for `sjoin` and `sjoin_pygeos` to fully optimize the `pygeos` implementation and have cleaner logic.

Test script:
```python3
import geopandas
import pandas as pd
from timeit import timeit
from geopandas import sjoin, read_file
import numpy as np

def import_dfs(data_len):
    world_path = geopandas.datasets.get_path("naturalearth_lowres")
    cities_path = geopandas.datasets.get_path("naturalearth_cities")
    countries = read_file(world_path)[["geometry", "name"]]
    countries = countries.rename(columns={"name": "country"})
    cities = read_file(cities_path)
    countries = countries.sample(n=data_len, replace=True, axis=0, random_state=1)
    cities = cities.sample(n=data_len, replace=True, axis=0, random_state=1)
    return cities, countries

def sjoin_test(data_len, use_pygeos):
    if hasattr(geopandas._compat, "set_use_pygeos"):
        # for easy testing of different branches
        geopandas._compat.set_use_pygeos(use_pygeos)
    cities, countries = import_dfs(data_len)
    t = timeit(lambda: sjoin(left_df=cities, right_df=countries), number=3)
    res = sjoin(left_df=cities, right_df=countries)
    return t, res


data_lens = [1, 10, 100, 1000]
times_pygeos = []
times_old = []
for data_len in data_lens:
    t_pygeos, res_pygeos = sjoin_test(data_len, use_pygeos=True)
    t_old, res_old = sjoin_test(data_len, use_pygeos=False)
    times_pygeos.append(t_pygeos)
    times_old.append(t_old)
    np.testing.assert_array_equal(
        # drop geometries because they will be different types
        res_pygeos.drop("geometry", axis=1),
        res_old.drop("geometry", axis=1),
    )

results = pd.DataFrame({
    "data_len": data_lens,
    "pygeos_time": times_pygeos,
    "old_time": times_old,
})
print(results)
```