@adriangb 2020-03-14
I decided to explore a cleaner way to implement this, as described above. The branch is on my repo, [adriangb/geopandas/pygeos-sindex](/adriangb/geopandas/tree/pygeos-sindex)

I was able to make a wrapper around `pygeos.STRTree` that emulates [Toblerity/rtree](/Toblerity/rtree). However, the implementation is not the cleanest. It can probably be cleaned up _some_, but while writing I was thinking that since `rtree` is an optional dependency and `pygeos` is going to be an optional dependency, it would be best to take the opportunity to remove `rtree` and use only `pygeos`. This would break the current API:

1. The `objects` param would be dropped from `sindex` methods.
2. Empty geometries will count towards the length of the `sindex` (but won't return results, I think).
3. If we want an even more streamlined implementation, I would phase out support of tuples of bounds as an input in favor of only accepting geometry objects. This might break some more things, but I think it makes sense to only accept one type. It could be a warning for version `0.x`.

Having a simple clean interface will become especially important once pygeos/pygeos#108 is implemented.